cmip6_data: cast masked 3D regrid output to float32 (halve resident state; fix assemble OOM)
Run 9zlhc cleared downloads + concat (masked 3D regrid complete) then OOMed in
assemble. Cause: xesmf upcasts the regrid to float64, so the pre-allocated
5-field 3D state was ~17GB (not ~8.4GB); _masked_regrid_3d_state itself fit but
assemble's fill-copies + thickness tipped the 32Gi pod over. Cast the pre-alloc
buffers to float32 (CMIP6 data + the zarr are float32 anyway) -> masked state
~8.4GB, assemble peak ~17GB. Correct + memory fix.

+++ scripts/cmip6_data/process_esgf.py
@@ -660,9 +660,13 @@ def _masked_regrid_3d_state(
         for v, da in seg_regridded.items():
             da = da.load()
             if v not in full_arrays:
-                full_arrays[v] = np.empty((n_time, *da.shape[1:]), dtype=da.dtype)
+                # float32: xesmf upcasts to float64, which would double the
+                # resident 3D state (5 fields x ~3.4GB) and OOM the pod during
+                # assembly. CMIP6 data is float32 and the pipeline stores
+                # float32, so cast here for both memory and storage fidelity.
+                full_arrays[v] = np.empty((n_time, *da.shape[1:]), dtype=np.float32)
                 template[v] = da  # for non-time coords/dims/attrs
-            full_arrays[v][sl] = da.values
+            full_arrays[v][sl] = da.values  # assignment casts to float32
         seg_valid = seg_valid.load()
         if valid_full is None:
             valid_full = np.empty((n_time, *seg_valid.shape[1:]), dtype=seg_valid.dtype)
